abstraction of GET Thresholds
