+++ src/humidity/humidity.repository.ts
@@ -3,6 +3,7 @@ import { PrismaService } from "nestjs-prisma";
 import { IntervalQuery } from "../shared/interval-query";
 import { NewThresholdDTO } from "../shared/newThresholdDTO";
 import { DataType } from "@prisma/client";
+import { getDatapointThresholds } from "../utils/thresholdQueryUtils";
 
 @Injectable()
 export class HumidityRepository {
@@ -33,12 +34,8 @@ export class HumidityRepository {
     });
   }
 
-  getDataPointThresholds() {
-    return this.prisma.thresholds.findUnique({
-      where: {
-        dataType: DataType.HUMIDITY,
-      },
-    });
+  getDatapointThresholds() {
+    return getDatapointThresholds(DataType.HUMIDITY, this.prisma);
   }
 
   postThresholdRequest(newThreshold: NewThresholdDTO) {
