+++ src/temperature/temperature.repository.ts
@@ -1,8 +1,9 @@
 import { Injectable } from "@nestjs/common";
 import { PrismaService } from "nestjs-prisma";
 import { IntervalQuery } from "../shared/interval-query";
-import { NewThresholdDTO } from "../shared/newThresholdDTO";
 import { DataType } from "@prisma/client";
+import { getDatapointThresholds } from "../utils/thresholdQueryUtils";
+import { NewThresholdDTO } from "../shared/newThresholdDTO";
 
 @Injectable()
 export class TemperatureRepository {
@@ -33,12 +34,8 @@ export class TemperatureRepository {
     });
   }
 
-  getDataPointThresholds() {
-    return this.prisma.thresholds.findUnique({
-      where: {
-        dataType: DataType.TEMPERATURE,
-      },
-    });
+  getDatapointThresholds() {
+    return getDatapointThresholds(DataType.TEMPERATURE, this.prisma);
   }
 
   postThresholdRequest(newThreshold: NewThresholdDTO) {
