+++ src/temperature/temperature.service.ts
@@ -16,7 +16,7 @@ export class TemperatureService {
   }
 
   getDataPointThresholds() {
-    return this.temperatureRepository.getDataPointThresholds();
+    return this.temperatureRepository.getDatapointThresholds();
   }
 
   postThresholdRequest(newThreshold: NewThresholdDTO) {
